# src/main/java/com/br/br/entity/Realm.java
//package com.br.br.entity;
//
//import lombok.Data;
//
//import javax.persistence.Entity;
//
//@Data
//@Entity
//public class Realm {
//
//
//    private Integer id = null;
//
//
//    private String name = null;
//
//
//    private String description = null;
//
//
//    private String key = null;
//
//
//}